add improved plotting submodule - move plot.py - add flopy plotting module (plot/flopy.py) - move all flopy style plotting into flopy.py - move plotting util functions to plotutil.py - add init
