+++ nlmod/plot/__init__.py
@@ -1,0 +1,2 @@
+from . import flopy
+from .plot import *
